projects/amazon-webscraper/requirements.txt
Flask
requests
beautifulsoup4
mysql-connector-python
python-dotenv